pyproject.toml
[project]
name = "oj-mcp-server"
version = "0.1.0"
description = "MCP server for Orange Juice Online Judge"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "requests",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
oj-mcp-server = "oj_mcp_server.server:main"
